Add systemd 258 compatibility
Systemd 258 recommends using `ACTION!="remove"` instead of the now
non-working `ACTION=="add"`. Although we've only applied this to the
Stadia device rules so far, it's good practice to set it for every rule.
Without this match, rules are applied to every state change, including
`remove`. Why would we grant access to a removed device?

This should also work on older system versions.

https://github.com/systemd/systemd/blob/v258/NEWS#L22-L32